+++ docs/CHANGES.rst
@@ -4,15 +4,22 @@ Changelog
 2.0.9 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Added locking behaviors on types.
+  [sgeulette]
+- Added MailingLoopTemplate type and mailing_loop_template field on ConfigurablePODTemplate.
+  [sgeulette]
+- Added 'mailing-loop-persistent-document-generation' view to manage mailing loop generation
+  [sgeulette]
+- Added helper method to manage context
+  [sgeulette]
+- Added helper method to check if mailed data have to be replaced during rendering
+  [sgeulette]
 
 2.0.8 (2017-08-02)
 ------------------
 
 - Add default value for 'pod_template' and 'output_format' attributes of the generation view.
   [sdelcourt]
-
 
 2.0.7 (2017-07-25)
 ------------------